Added cpiofile.py module
The plan is to make this mimic the tarfile module, except for cpio
files (the only currently planned target is the ASCII cpio format).
Once this is done, the pax utility can be implemented with full
functionality.

Also included the test.cpio file to work with when testing the
module. It is an archive of a .obj file and a .mtl file, both
relatively small in size.